cli: implement lion's share of 'run' command.
Serialization and deserialization glue still needed, and more test rigor for the top level stuff, but it's close.

Added constructor to chroot implementation, which was badly needed.  Our choices in rounding off edges in that area is pretty sketch -- rio is still doing WAY too much flying by env vars, under that deceptively-simple construct for stitch.Assembler -- but it's working well enough so far; we can refactor and refine that area later.

Signed-off-by: Eric Myhre <hash@exultant.us>