all: update for formula/context split.
That went surprisingly gracefully.

Also included: update to errcat.  I would've rather separated it, but I got in the odd situation of having updated the *other* upstream libraries in an inconsistent order, and thus got a diamond problem.  So, it's just rolled in here.  It's only like two extra lines of diff, fortunately.

Signed-off-by: Eric Myhre <hash@exultant.us>