executor/cradle: userinfo, and default settings.
This API bump brings in userinfo, and the ability to specify UIDs, GIDs, usernames, and so on.

The default UID if unspecified is 1000 again, not zero, as it should be.  (Sure, we've got containers, "it should be fine".  But screw that logic: safe defaults are safe, safe is better, and if you have programs that can't run as arbitrary UIDs, *it's a bug* and you should at the very least know about it.)

The username and homedir functionality are silent as of yet.  They will become impactful when we procede to implement the filesystem setup parts of the cradle logic, which is coming up shortly.

Signed-off-by: Eric Myhre <hash@exultant.us>