executor,libs: rio update, and can now check more environment before launching executor.
That whole application-level symlink traversal sandbox, though a lot of work, is now done, and makes new error path tests like this fly right.  (It's quiet, but we're already exercising some of the edge cases that previously would've failed: the busybash fixture `/bin/echo` is a symlink!  More tests with more explicit purpose and coverage could, of course, still be used.)

Signed-off-by: Eric Myhre <hash@exultant.us>