executor/cradle: first pass on filesystem setup.
This is *incorrect*: note how we had to add `cradle=disable` to a test in order to keep it green -- we're not doing correct mtime maint in these functions.  It's also semantically incorrect: repairing permissions if the dirs already do exist is *not optional*, and in fact this will show up in you try to actually *use* the tools with our fixture item... it has the /task dir and it's owned by root!  Heh.

Signed-off-by: Eric Myhre <hash@exultant.us>