example of pipeline with hitch releases.
This is the *most primitive* thing that can *arguably work*: bash heredocs galore.  And yet: *it works*.  It is an outright design goal to keep something this utterly curmudgeonly bashy and simplistic possible.

Mind, this will also fail in a half a dozen ways without additional env setup that I'm currently handwaving on (having a hitch binary on your path; having the bootstrap artifact already cached; soon we're going to lean on jq... etc)... this is just a POC as yet.

Signed-off-by: Eric Myhre <hash@exultant.us>