executor/impl/chroot: of course, fix error branch.
In the quest for brevity and the swift strokes of a refactor, what did we forget?  An error branch, of course.  It's almost like a language should have BUILT IN TOOLS to help you detect these mistakes since they're utterly critical and the situations occur daily, nay, hourly, nay, each and every minute that you write code.

Signed-off-by: Eric Myhre <hash@exultant.us>