executor: reusable input forwarder.
I'm... not entirely sure about all the depths of whether I should prefer to use StdinPipe() vs setting StdIn to a reader.  The godocs seem to indicate that the latter involves the *runtime* spawning a goroutine for you; it's not clear from the docs (and I haven't bothered to read source deeply enough) to see if both paths involve allocating a fd for the pipe.  At any rate, we're going the road we A) tbh were going before and B) conserves a duplicate goroutine (since we need one to bounce the channel, regardless).

Signed-off-by: Eric Myhre <hash@exultant.us>