executor: reusable 'WithFilesystem'.
This is a dubious simplification -- the lines-of-code count *increases* in this commit -- but it should pay itself back down when we add a runc executor which needs roughly the exact same work done for everything except the exec itself... which is the abstraction level this tries to reach.

As you can see, most parts of each half require... well, *most* of the arguements :(  There are a few wins though: it's now clear that the filesystem setup never needs the 'input'; and the run itself never needs the 'formulaCtx' (!!!! as deity intended).

Signed-off-by: Eric Myhre <hash@exultant.us>