cmd: fix issue with hang from executor setup err.
Always be careful setting up channels, blocking on their complete servicing, and then not being 100% sure those channels will be closed.

In this case, the problem was from setting up such a channel, doing *more* unrelated setup, and *then* starting the process which is in charge of the channel closure.  This refactor gets that stuff in the middle back out again.

Also, put log printing into the twerk executor, because there's not really a reason it was missing there (though we may still choose to have those decorate somewhat differently; we'll see).

Signed-off-by: Eric Myhre <hash@exultant.us>