fling: add install-deps command, and include in default.
The install-deps command just covers the rio build right now.

Arguably, the 'test' command might want to shell out to this too, but this is pretty much how we did it before too.  (It's a tad worrying to put TOO many things in Every Single build step when they can't quite reduce to instantaneous no-ops when already done.)

Signed-off-by: Eric Myhre <hash@exultant.us>