delete goad; switch CI to fling.
Add GOBIN to PATH in the preamble include so the `rio` binary is seen in tests.  Maaaybe not a good hack.  (Irritating because the more deltas in these fling files, the sadder I am by default; also dubious because... is this really how we want our plugins to work?  No, we're probably being sloppy and this should be handled in more code in our real code.)  Makes tests continue to be green though, so we can call this script migration complete; THEN lift our eyes to making this better.

Signed-off-by: Eric Myhre <hash@exultant.us>