cmd: fix line break slippage in output.
This is one of those irritating things where the technical correctness is at no point *wrong*, because the outputs are going on different channels -- stdout and stderr -- but nonetheless the visual output for a human at a terminal was coming out wrong because the streams and their bufferings are not synchronized.

Anyway, now it's fixed, by bothering to do a short buffer in memory.  Yay for flush control.

Signed-off-by: Eric Myhre <hash@exultant.us>