fling: add fetch-plugins command; mirror cache; gitignore all this.
This mirror dir should be regenerateable from public sources; running fetch-plugins once should make them available in this CAS way, so you can switch branches with different libs, fetch more, switch back, cache is still good, etc.

You'll always have to run `fling fetch-plugins` once after cloning the repo (or switching to a branch with different plugins or versions of them), but after that you should be good to go offline.

Right now we'll just have the runc binary plugin.  This is fetched from an existing, known version rather than built as part of the repeatr build recursion because it's a Big, Hairy process, requring a C compiler, and all sorts of fuckery; it's slow, doesn't memoize well of its own accord (though of course it does in a formula), and so on.  This binary comes from a formula in https://github.com/polydawn/formulary ; we've been using it for quite a while already.  Someday, we'll replace it with a newer one, and by then hopefully also have hitch replay records in place so I can link directly to unambiguous repro instructions.  That's down the road a tad, yet :3  But close!

Signed-off-by: Eric Myhre <hash@exultant.us>