cmd: very tentative first impl of builtin memoization.
Just the reader part.  Memo files are not yet generated.

Note the several "// REVIEW" comments.

Previous generations of repeatr did *no* built-in memoization; we built that feature at a higher layer.  Moving it *in* to repeatr itself seems like it should be an improvement... but there are of course questions about what exactly the "sensible" defaults should be with such a feature.

Signed-off-by: Eric Myhre <hash@exultant.us>