cmd: memoization saving.
and small fix to deserialization.  (thought I had a bug in refmt there, for a moment, but stdlib json does the same thing.  And I guess it makes sense once you refresh yourself on the concepts again for a moment: the pointer *itself* isn't addressable in that scenario, so of course you can't initialize it for the caller.)

feature works now (...!) modulo the lack of configuration option to enable it -- still a hardcoded nil there.

and we start to need a test framework for stuff up here.  Which is easier said than done: we're starting to sprout some pretty sizable "saga" situations, and we need to be able write tests for these both clearly and concisely.

Signed-off-by: Eric Myhre <hash@exultant.us>