Add transitive deps of rio.
This is... an odd commit.
**We don't need these deps to build repeatr.**
Repeatr uses some library packages from Rio, but does
*not* in its own build depend on any of these packages
(git, ssh, etc).  Repeatr uses Rio via IPC in part
specifically to not have these dependencies.

But it's kind of crap if our build deps for a *working*
repeatr, which does depend on having that rio binary to exec,
starts depending on... rio/repeatr for a build pipeline.
*In every other way*, having a rio/repeatr powered build pipeline
would be *the best possible solution* for these needs.

But we'd rather avoid a bootstrap dependency cycle.

So.  Here are all the Rio transitive deps.
So that we can keep having the Repeatr build script also
build Rio from this same repo.

Womp womp.

(The good news of course is... it's git submodules.
So there's no permanent historical baggage or scaling issues
in this repo, like there would be if we were doing something
foolish like vendoring.  And if you *are* building
things in rio/repeatr pipelines, your submodules will even
cache individually... so there's *no* overhead at all ;))

Signed-off-by: Eric Myhre <hash@exultant.us>