Implement memoization as a RunFunc decorator.
This is a way more excellent implementation than before.

I was pushed to look into this while trying to factor out the relevant parts of
the main.Run method for the batching mode in progress on another branch.
Getting the memoization extracted cleans up a *lot* of the guff.

The contract of who closes monitor chans has been updated.  It's the caller.
This turns out a lot nicer: it removes almost all the conditionals that used to
exist about the potential of nil monitors.  This was forced by the issue of how
the memoizing decorator could possibly log correctly if the executor it
decorates might have already closed the monitor message channel.  This felt
moderately scary, but I think our regression test coverage for this is actually
good enough that we can be pretty confident it's fine.

Signed-off-by: Eric Myhre <hash@exultant.us>