Prettier printing of output and logs.
Leaving room for plugging in more alterative modes later.
We'll want a json API mode of course; also probably less colorful modes.

Momentarily regressing on the pretty print of the runrecord object json;
refmt is going to have that feature implemented in a more proper way in
a few shakes and I'd rather do it that way.

Signed-off-by: Eric Myhre <hash@exultant.us>