Buffer incomplete lines properly when colorizing output.
This means things can get delayed if the linebreak comes long after the start
of the line... but there's really no other sensible thing to do when we have
the requirement of emitting full lines and trying to keep things as lossless
as we possibly can.

The previous behavior would flush things somewhat over-eagerly and with extra
line breaks.  This sometimes looked... well, terrible.

Signed-off-by: Eric Myhre <hash@exultant.us>