examples: restrict PATH to GOBIN.
If you have repeatr installed globally, we wouldn't want your tests to
make the mistake of using that one when there's no from-source build.

Probably it would be nice if the package init made sure the whole
system is built, but that's a bit of an stretch at the moment.
(Are we really happy with the complexity level of the fling scripts?
Not particularly.  Do we have a clear intuition of how to make it
better, given that we actually *do* need to build *two* binaries in
this project plus fetch and unpack some externally built and pinned
tools? Also no. :/ )

Signed-off-by: Eric Myhre <hash@exultant.us>