examples: isolate cache dirs, use them in groups.
This lets us assert on stderr logs correctly, with both cold and hot
caches.

I'm giving up all impression that these testcase files can be
automatically discovered and used in any ol' order.  It's a nice idea,
but doesn't survive well in contact with reality: in addition to being
a bunch of extra code, for your trouble you'll get... even less
relevant line numbers in the Go test log output.  Nah.
You can still run any of the formulas independently in your shell,
though -- only the stderr fixtures aren't guaranteed to match.

Further on... we're going to have issues with the ordering of log
statements for transmats, both on pack and unpack, because those *are*
actually parallelized and are free to complete in any order.
I don't yet have a plan on how to deal with that; it's possible we just
*won't* check stderr on formulas with plural inputs or outputs... or
at least, not with this kind of test.  Or, we might research something
that can detect permutable lines, and do another "pave" func that
sorts them when they're found in clusters; that might do a lot.
Regardless: that's future work.

Signed-off-by: Eric Myhre <hash@exultant.us>