Skip tests that require root if power unavailable.
There's a somewhat cleverer version of this in one of the old branches
(from back before the v200 rewrite) which actually examines
capabilities bits; it'd be nice to port that forward again, but I'd
like to give it a serious think and maybe try to get some very meta
tests around those conditions themselves.

So this is back to the most basic practical thing that works.

Signed-off-by: Eric Myhre <hash@exultant.us>