experimental gvisor support.
Requires some additional setup: you have to provide your own gvisor binary.
The gvisor team is still only producing nightly builds themselves, and no
official releases, so I'm loathe to arbitrarily pick a nightly and latch onto
it and have to rehost it for eternity.

That said, things work.

Most of this code is the same: gvisor really is a drop-in for runc.
I stripped out some extra config that runc needed and seems to be unimpactful
for gvisor.  Other than that, this code is quite cookie cutter... so far.
I'm creating all the packages and even allowing some duplication simply so it's
clear where we would make more significant modifications in the future.

This does *not* give us the ability to immediately running repeatr as a whole
in a rootless low-priv fashion, unfortunately.  We still need escalated privs
to assemble our whole stitch-up of mounts from the outside of the container.
It's possible that we could try to move some of the features for this into
gvisor itself; they seem open to the concept, and doing so would let us get
truly rootless operation.  [*]

(Of course, no, that still wouldn't be enough -- one would need to be able to
chown files, etc, on the host, in order to run many forms of `rio unpack`.
So... there's more than a little work between the present and this goal.
Someday it might even be "easiest" to write a gofer filesystem provider plugin
to gvisor.  That's... a whole ball of wax, though.)

gvisor's isolation really is great.  Once they get real releases rather than
only nightlies, it's very likely that I'll switch the default executor over
to gvisor.

Signed-off-by: Eric Myhre <hash@exultant.us>