Shared API update.  This is kind of massive.
We're jumping over to the "redux" branch in go-timeless-api, which has
been brewing for quite a while now, and mostly under the auspices of
"do what stellar/pipelining needs"... and after a few months of that,
it's time to sync up all the timeless stack projects on the result.

Which has involved fixing a lot of things in the new "redux" branch to
be less specifically captured by the interests of the pipelining layer
they were developed with as a primary concern, of course.

So anyway.

Repeatr functionality overall hasn't changed.  Like, _at all_.

The entire formulaContext system for URLs: still here.

The entire formula schema: still here.

There's a slight change to filters.  Those are now specified as
strings (think: like "options" for mounts in /etc/fstab).  This might
cause any formulas which used output filters to need some minor
updates.  However, formulas which use non-default filters are rare in
the first place, so this shouldn't be very disruptive.

Support for "batch"/basting mode is dropped, completely.  That was an
experiment, and clearly marked as such; you were warned.
The Stellar project contains tools for handling the the (much, much,
MUCH) better system, "modules", that flowered out of that experiment.
Repeatr (now, again) no longer has ANY code at all which handles the
idea of more than one formula as a time -- and that's as it should be.
Higher level tools will handle the higher level concepts.

Symptomatically, all the ordering code in the batch package disappears.
It's actually already replicated (and grown) in the go-timeless-api
"redux" branch that we're now getting repeatr up-to-date with, since
it's used in the "module" system that's replaced basting.

Some error codes have been been bumped up and around.  Repeatr errors
are now in the 40's.  JobUnsuccessful remains at 32.  This might affect
you if you wrote scripts that check numerical process exits.  Mostly,
this shouldn't have much effect, however; if you really needed
a lot of scripting around specific error handling, you probably should
be using the more powerful and explicit json API modes anyway, and
those error codes have not been changed.

Some code which was previously in rio//stitch has disappeared there and
reappeared in repeatr//executor/mixins.  This is stuff that handled
flipping formula inputs and outputs into individual rio operations, and
since it spent so much time referencing formulas, should in hindsight
never have been in Rio in the first place.

The 'printing.go' file from repeatr//cmd/repeatr has disappeared.
We moved that functionality to go-timeless-api//repeatr/fmt some while
ago, since Stellar has the same printers.  I think there's More Work To
Do on making those composable *nicely*, but that's where they live now.

Other than that, a lot of the changes here are purely mechanical and
minor internal nomenclature changes, fairly literally sed:

- sedr 's/api.FormulaContext/repeatr.FormulaContext/g'
- sedr 's/api.RunRecord/api.FormulaRunRecord/g'
- sedr 's#"go.polydawn.net/rio/client"#"go.polydawn.net/go-timeless-api/rio/client/exec"#'
- sedr 's/rioexecclient/rioclient/g'
- sedr 's/OutputSpec/FormulaOutputSpec/g'

Overall these reflect that Rio client code moved; and that many of
the API types we use in Repeatr are now prefixed with "Formula" to
make them distinct from similar types that describe an
"Operation" (a proto-formula, at the new "module" level).

Oh, and since we're making a big jump in refmt library version, we're
also getting to use "unions" from that system, which makes a bunch of
the RPC Event/Output/Result code significantly cleaner.  Also,
"omitempty" has been applied for a lot more things.

Relatedly, a lot of the code around Monitors has gotten a bit simpler.
It both no longer needs wrapper types (because of refmt union support),
and now has a Send helper method from go-timeless-api that lets us skip
a lot of checks for if we have an initialized channel.  Terser.

And that's about it.

Big internal changes.  Minimal external changes.  And we're now synced
up with the "redux" branch on go-timeless-api, which is a major step
forward; we'll probably be making that branch the master branch next.

Signed-off-by: Eric Myhre <hash@exultant.us>