Bump go-timeless-api; fix "filters" backcompat.
We were considering making breaking changes to how filters are
specified, since we *did* also just improve the granularity with which
they can be configured.  However, on further review, there's just... no
real reason to make an API break, because the previous definitions
actually still gracefully "do the right thing".

So, we've made support in go-timeless-api for both stringy and object
style serializations of filters, and named the key in output specs as
"filters" (plural) (again, as before), and it's serialized in the
object style.  Overall, this is now a non-breaking change.  Tada.

Signed-off-by: Eric Myhre <hash@exultant.us>