Include golang.org/x/sys in submodules.
This is used if you build Rio from within this source tree...
which is what the 'fling' script does.

(If you build Rio separately and just bring it into your $PATH, this
library isn't necessary to have here.)

I missed this because none of the modules that build the whole
stack as a whole actually do this thing that the 'fling' script does.

It may be time for a little clean up of some of this dev stuff.